requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-genai>=0.1.0
psycopg2-binary
python-dotenv
